requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
pydantic-core==2.18.4
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
